fix: revisi
