+++ src/components/attendance/admin/event-management/AddEventModal.tsx
@@ -99,6 +99,7 @@ export const AddEventModal = ({ addEvent, day }: AddDayModalProps) => {
         onClick={onOpen}
         mt={{ base: 'none', lg: '1em' }}
         w={{ base: '100%', lg: '8em' }}
+        h='2em'
       >
         Tambah Event
       </Button>
