+++ src/components/attendance/mentor/mentor-recap/MentorRecap.tsx
@@ -153,7 +153,8 @@ export const MentorRecap = ({ dayId }: MentorRecapProps) => {
           <>
             <Button
               variant='mono-outline'
-              w='8em'
+              w={{ base: '30%', lg: '6em' }}
+              h='2em'
               onClick={onEditingRowPerPageOpen}
             >
               {rowPerPage}
@@ -273,7 +274,8 @@ export const MentorRecap = ({ dayId }: MentorRecapProps) => {
         <Flex justifyContent={{ base: 'space-between', lg: 'right' }}>
           <Button
             variant='mono-outline'
-            w={{ base: '30%', lg: '6em' }}
+            w={{ base: '30%', lg: '4em' }}
+            h='2em'
             mr='1em'
             onClick={prevPage}
           >
@@ -285,16 +287,16 @@ export const MentorRecap = ({ dayId }: MentorRecapProps) => {
               borderRadius='12px'
               color='gray.600'
               w={{ base: '30%', lg: '6em' }}
-              h='2.5em'
+              h='2em'
             >
               {`${page}`}
             </MenuButton>
             <MenuList border='1px solid gray' p='1em'>
-              <Flex>
+              <Flex alignItems='center'>
                 <Input value={jumpInput} onChange={jumpChangeHandler} />
                 <Button
                   variant='mono-outline'
-                  w='8em'
+                  w={{ base: '30%', lg: '4em' }}
                   ml='1em'
                   onClick={jumpToPage}
                 >
@@ -305,7 +307,8 @@ export const MentorRecap = ({ dayId }: MentorRecapProps) => {
           </Menu>
           <Button
             variant='mono-outline'
-            w={{ base: '30%', lg: '6em' }}
+            w={{ base: '30%', lg: '4em' }}
+            h='2em'
             ml='1em'
             onClick={nextPage}
           >
